@@ -1,5 +1,5 @@
-import AsyncAirtable = require('../asyncAirtable');
-import { AirtableRecord } from '../@types';
+import { AsyncAirtable } from '../asyncAirtable';
+import { AirtableRecord } from '../types';
 import { config } from 'dotenv';
 config();
 const asyncAirtable = new AsyncAirtable(
@@ -8,7 +8,7 @@ const asyncAirtable = new AsyncAirtable(
 );
 let created: AirtableRecord;
 describe('.createRecord', () => {
-  test('should create a new entry in the table with the given fields', async (done) => {
+  test('should create a new entry in the table with the given fields', async () => {
     const result = await asyncAirtable.createRecord(
       process.env.AIRTABLE_TABLE || '',
       JSON.parse(process.env.NEW_RECORD || ''),
@@ -21,10 +21,9 @@ describe('.createRecord', () => {
     expect(result.createdTime).toBeDefined();
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     created = result;
-    done();
   });
 
-  test('should be able to find the record by the id after creation', async (done) => {
+  test('should be able to find the record by the id after creation', async () => {
     const result = await asyncAirtable.find(
       process.env.AIRTABLE_TABLE || '',
       created.id,
@@ -37,45 +36,40 @@ describe('.createRecord', () => {
     expect(result.createdTime).toBeDefined();
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     expect(JSON.stringify(result)).toEqual(JSON.stringify(created));
-    done();
   });
 
-  test('should throw an error if you do not pass a table', async (done) => {
+  test('should throw an error if you do not pass a table', async () => {
     // @ts-ignore
     await expect(asyncAirtable.createRecord()).rejects.toThrowError(
       'Argument "table" is required',
     );
-    done();
   });
 
-  test('should throw an error if you do not pass a record', async (done) => {
+  test('should throw an error if you do not pass a record', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.createRecord(process.env.AIRTABLE_TABLE || ''),
     ).rejects.toThrowError('Argument "record" is required');
-    done();
   });
 
-  test('should throw an error if pass a field that does not exist', async (done) => {
+  test('should throw an error if pass a field that does not exist', async () => {
     await expect(
       asyncAirtable.createRecord(process.env.AIRTABLE_TABLE || '', {
         gringle: 'grangle',
       }),
     ).rejects.toThrowError(/UNKNOWN_FIELD_NAME/g);
-    done();
   });
 
-  test('should throw an error if pass a field with the incorrect data type', async (done) => {
+  test('should throw an error if pass a field with the incorrect data type', async () => {
     await expect(
       asyncAirtable.createRecord(process.env.AIRTABLE_TABLE || '', {
         ...JSON.parse(process.env.NEW_RECORD || ''),
         value: 'nope',
       }),
     ).rejects.toThrowError(/INVALID_VALUE_FOR_COLUMN/g);
-    done();
   });
 
-  test('should throw an error if pass the table argument with an incorrect data type', async (done) => {
+  test('should throw an error if pass the table argument with an incorrect data type', async () => {
     await expect(
       asyncAirtable.createRecord(
         // @ts-ignore
@@ -83,37 +77,12 @@ describe('.createRecord', () => {
         JSON.parse(process.env.NEW_RECORD || ''),
       ),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
   });
 
-  test('should throw an error if pass the record argument with an incorrect data type', async (done) => {
+  test('should throw an error if pass the record argument with an incorrect data type', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.createRecord(process.env.AIRTABLE_TABLE || '', 10),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
-  });
-
-  test('should retry if rate limited', async (done) => {
-    let results = [];
-    for (let i = 0; i < parseInt(process.env.REQ_COUNT || ''); i++) {
-      results.push(
-        asyncAirtable.createRecord(
-          process.env.AIRTABLE_TABLE || '',
-          JSON.parse(process.env.NEW_RECORD || ''),
-        ),
-      );
-    }
-    const data: AirtableRecord[] = await Promise.all(results);
-    data.forEach((result) => {
-      expect(result).toBeDefined();
-      expect(typeof result).toBe('object');
-      expect(Object.keys(result).length).toBeGreaterThan(0);
-      expect(result).toHaveProperty('id');
-      expect(result).toHaveProperty('fields');
-      expect(result).toHaveProperty('createdTime');
-      expect(Object.keys(result.fields).length).toBeGreaterThan(0);
-    });
-    done();
   });
 });
