@@ -1,5 +1,5 @@
-import AsyncAirtable = require('../asyncAirtable');
-import { AirtableRecord } from '../@types';
+import { AsyncAirtable } from '../asyncAirtable';
+import { AirtableRecord } from '../types';
 import { config } from 'dotenv';
 config();
 const asyncAirtable = new AsyncAirtable(
@@ -10,7 +10,7 @@ let firstResult: AirtableRecord;
 let secondResult: AirtableRecord;
 let compare: AirtableRecord;
 describe('.find', () => {
-  beforeAll(async (done) => {
+  beforeAll(async () => {
     const testResult = await asyncAirtable.select(
       process.env.AIRTABLE_TABLE || '',
       {
@@ -19,10 +19,9 @@ describe('.find', () => {
     );
     firstResult = testResult[0];
     secondResult = testResult[1];
-    done();
   });
 
-  test('should find a specific record by Airtable ID', async (done) => {
+  test('should find a specific record by Airtable ID', async () => {
     const result = await asyncAirtable.find(
       process.env.AIRTABLE_TABLE || '',
       firstResult.id,
@@ -35,10 +34,9 @@ describe('.find', () => {
     expect(result.createdTime).toBeDefined();
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     compare = result;
-    done();
   });
 
-  test('should find a different specific record by Airtable ID', async (done) => {
+  test('should find a different specific record by Airtable ID', async () => {
     const result = await asyncAirtable.find(
       process.env.AIRTABLE_TABLE || '',
       secondResult.id,
@@ -51,72 +49,45 @@ describe('.find', () => {
     expect(result.createdTime).toBeDefined();
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     expect(JSON.stringify(result)).not.toEqual(JSON.stringify(compare));
-    done();
   });
 
-  test('should throw an error if you do not pass a table', async (done) => {
+  test('should throw an error if you do not pass a table', async () => {
     // @ts-ignore
     await expect(asyncAirtable.find()).rejects.toThrowError(
       'Argument "table" is required',
     );
-    done();
   });
 
-  test('should throw an error if the table does not exist', async (done) => {
+  test('should throw an error if the table does not exist', async () => {
     await expect(
       asyncAirtable.find('doesnotexist', firstResult.id),
     ).rejects.toThrowError(/"TABLE_NOT_FOUND"/g);
-    done();
   });
 
-  test('should throw an error if you pass an incorrect data type for table', async (done) => {
+  test('should throw an error if you pass an incorrect data type for table', async () => {
     // @ts-ignore
     await expect(asyncAirtable.find(10)).rejects.toThrowError(
       /Incorrect data type/g,
     );
-    done();
   });
 
-  test('should throw an error if you do not pass an id', async (done) => {
+  test('should throw an error if you do not pass an id', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.find(process.env.AIRTABLE_TABLE || ''),
     ).rejects.toThrowError('Argument "id" is required');
-    done();
   });
 
-  test('should throw an error if the id does not exist', async (done) => {
+  test('should throw an error if the id does not exist', async () => {
     await expect(
       asyncAirtable.find(process.env.AIRTABLE_TABLE || '', 'doesnotexist'),
     ).rejects.toThrowError(/"NOT_FOUND"/g);
-    done();
   });
 
-  test('should throw an error if you pass an incorrect data type for table', async (done) => {
+  test('should throw an error if you pass an incorrect data type for table', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.find(process.env.AIRTABLE_TABLE || '', 10),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
-  });
-
-  test('should retry if rate limited', async (done) => {
-    let results = [];
-    for (let i = 0; i < parseInt(process.env.REQ_COUNT || ''); i++) {
-      results.push(
-        asyncAirtable.find(process.env.AIRTABLE_TABLE || '', firstResult.id),
-      );
-    }
-    const data: AirtableRecord[] = await Promise.all(results);
-    data.forEach((result) => {
-      expect(result).toBeDefined();
-      expect(typeof result).toBe('object');
-      expect(Object.keys(result).length).toBeGreaterThan(0);
-      expect(result).toHaveProperty('id');
-      expect(result).toHaveProperty('fields');
-      expect(result).toHaveProperty('createdTime');
-      expect(Object.keys(result.fields).length).toBeGreaterThan(0);
-    });
-    done();
   });
 });
