@@ -1,5 +1,5 @@
-import AsyncAirtable = require('../asyncAirtable');
-import { AirtableRecord } from '../@types';
+import { AsyncAirtable } from '../asyncAirtable';
+import { AirtableRecord } from '../types';
 import { config } from 'dotenv';
 config();
 const asyncAirtable = new AsyncAirtable(
@@ -8,16 +8,15 @@ const asyncAirtable = new AsyncAirtable(
 );
 let initResult: AirtableRecord[];
 describe('.updateRecord', () => {
-  beforeAll(async (done) => {
+  beforeAll(async () => {
     initResult = await asyncAirtable.select(process.env.AIRTABLE_TABLE || '', {
-      maxRecords: 1,
+      maxRecords: 2,
       sort: [{ field: 'value', direction: 'desc' }],
       view: 'Grid view',
     });
-    done();
   });
 
-  test('should update a record with provided data', async (done) => {
+  test('should update a record with provided data', async () => {
     const result = await asyncAirtable.updateRecord(
       process.env.AIRTABLE_TABLE || '',
       {
@@ -34,10 +33,9 @@ describe('.updateRecord', () => {
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     expect(JSON.stringify(result)).not.toEqual(JSON.stringify(initResult[0]));
     initResult[0] = result;
-    done();
   });
 
-  test('should update a record and set unprovided field to null if you pass in the destructive arg', async (done) => {
+  test('should update a record and set unprovided field to null if you pass in the destructive arg', async () => {
     const result = await asyncAirtable.updateRecord(
       process.env.AIRTABLE_TABLE || '',
       {
@@ -55,26 +53,23 @@ describe('.updateRecord', () => {
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     expect(JSON.stringify(result)).not.toEqual(JSON.stringify(initResult[0]));
     expect(result).not.toHaveProperty('email');
-    done();
   });
 
-  test('should throw an error if you do not pass a table', async (done) => {
+  test('should throw an error if you do not pass a table', async () => {
     // @ts-ignore
     await expect(asyncAirtable.updateRecord()).rejects.toThrowError(
       'Argument "table" is required',
     );
-    done();
   });
 
-  test('should throw an error if you do not pass a record', async (done) => {
+  test('should throw an error if you do not pass a record', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.updateRecord(process.env.AIRTABLE_TABLE || ''),
     ).rejects.toThrowError('Argument "record" is required');
-    done();
   });
 
-  test('should throw an error if you pass a field that does not exist', async (done) => {
+  test('should throw an error if you pass a field that does not exist', async () => {
     await expect(
       asyncAirtable.updateRecord(process.env.AIRTABLE_TABLE || '', {
         id: initResult[0].id,
@@ -82,20 +77,18 @@ describe('.updateRecord', () => {
         fields: { gringle: 'grangle' },
       }),
     ).rejects.toThrowError(/UNKNOWN_FIELD_NAME/g);
-    done();
   });
 
-  test('should throw an error if you send an incorrect id', async (done) => {
+  test('should throw an error if you send an incorrect id', async () => {
     await expect(
       asyncAirtable.updateRecord(process.env.AIRTABLE_TABLE || '', {
         id: 'doesnotexist',
         fields: JSON.parse(process.env.UPDATE_RECORD || ''),
       }),
     ).rejects.toThrowError(/NOT_FOUND/g);
-    done();
   });
 
-  test('should throw an error if pass a field with the incorrect data type', async (done) => {
+  test('should throw an error if pass a field with the incorrect data type', async () => {
     await expect(
       asyncAirtable.updateRecord(process.env.AIRTABLE_TABLE || '', {
         id: initResult[0].id,
@@ -105,10 +98,9 @@ describe('.updateRecord', () => {
         },
       }),
     ).rejects.toThrowError(/INVALID_VALUE_FOR_COLUMN/g);
-    done();
   });
 
-  test('should throw an error if pass the table argument with an incorrect data type', async (done) => {
+  test('should throw an error if pass the table argument with an incorrect data type', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.updateRecord(10, {
@@ -116,38 +108,12 @@ describe('.updateRecord', () => {
         fields: JSON.parse(process.env.UPDATE_RECORD || ''),
       }),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
   });
 
-  test('should throw an error if pass the record argument with an incorrect data type', async (done) => {
+  test('should throw an error if pass the record argument with an incorrect data type', async () => {
     await expect(
       //@ts-ignore
       asyncAirtable.updateRecord(process.env.AIRTABLE_TABLE || '', 10),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
-  });
-
-  test('should if rate limited', async (done) => {
-    let results = [];
-    for (let i = 0; i < parseInt(process.env.REQ_COUNT || ''); i++) {
-      results.push(
-        asyncAirtable.updateRecord(process.env.AIRTABLE_TABLE || '', {
-          id: initResult[0].id,
-          fields: JSON.parse(process.env.UPDATE_RECORD || ''),
-        }),
-      );
-    }
-    const data: AirtableRecord[] = await Promise.all(results);
-    data.forEach((result) => {
-      expect(result).toBeDefined();
-      expect(typeof result).toBe('object');
-      expect(Object.keys(result).length).toBeGreaterThan(0);
-      expect(result.id).toBeDefined();
-      expect(result.fields).toBeDefined();
-      expect(result.createdTime).toBeDefined();
-      expect(Object.keys(result.fields).length).toBeGreaterThan(0);
-      expect(JSON.stringify(result)).not.toEqual(JSON.stringify(initResult[0]));
-    });
-    done();
   });
 });
