@@ -1,5 +1,5 @@
-import AsyncAirtable = require('../asyncAirtable');
-import { AirtableRecord } from '../@types';
+import { AsyncAirtable } from '../asyncAirtable';
+import { AirtableRecord } from '../types';
 import { config } from 'dotenv';
 config();
 const asyncAirtable = new AsyncAirtable(
@@ -8,17 +8,16 @@ const asyncAirtable = new AsyncAirtable(
 );
 let initResult: AirtableRecord[];
 describe('.upsertRecord', () => {
-  beforeAll(async (done) => {
+  beforeAll(async () => {
     initResult = await asyncAirtable.select(process.env.AIRTABLE_TABLE || '', {
       filterByFormula: "{title} = 'test-find'",
       maxRecords: 1,
       sort: [{ field: 'value', direction: 'asc' }],
       view: 'Grid view',
     });
-    done();
   });
 
-  test('should update a record with provided data if it exists', async (done) => {
+  test('should update a record with provided data if it exists', async () => {
     const result = await asyncAirtable.upsertRecord(
       process.env.AIRTABLE_TABLE || '',
       "{title} = 'test-upsert'",
@@ -33,10 +32,9 @@ describe('.upsertRecord', () => {
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     expect(JSON.stringify(result)).not.toEqual(JSON.stringify(initResult[0]));
     initResult[0] = result;
-    done();
   });
 
-  test('should update a record with provided data desctructively if it exists', async (done) => {
+  test('should update a record with provided data desctructively if it exists', async () => {
     const result = await asyncAirtable.upsertRecord(
       process.env.AIRTABLE_TABLE || '',
       "{title} = 'test-upsert'",
@@ -51,10 +49,9 @@ describe('.upsertRecord', () => {
     expect(result.createdTime).toBeDefined();
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     expect(result.fields).not.toHaveProperty('email');
-    done();
   });
 
-  test('should create a new record with provided data if one does not exist', async (done) => {
+  test('should create a new record with provided data if one does not exist', async () => {
     const result = await asyncAirtable.upsertRecord(
       process.env.AIRTABLE_TABLE || '',
       "{title} = 'test-test'",
@@ -68,26 +65,23 @@ describe('.upsertRecord', () => {
     expect(result.createdTime).toBeDefined();
     expect(Object.keys(result.fields).length).toBeGreaterThan(0);
     expect(result.id).not.toEqual(initResult[0].id);
-    done();
   });
 
-  test('should throw an error if you do not pass a table', async (done) => {
+  test('should throw an error if you do not pass a table', async () => {
     // @ts-ignore
     await expect(asyncAirtable.upsertRecord()).rejects.toThrowError(
       'Argument "table" is required',
     );
-    done();
   });
 
-  test('should throw an error if you do not pass a filterString', async (done) => {
+  test('should throw an error if you do not pass a filterString', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.upsertRecord(process.env.AIRTABLE_TABLE || ''),
     ).rejects.toThrowError('Argument "filterString" is required');
-    done();
   });
 
-  test('should throw an error if you do not pass a record', async (done) => {
+  test('should throw an error if you do not pass a record', async () => {
     await expect(
       // @ts-ignore
       asyncAirtable.upsertRecord(
@@ -95,10 +89,9 @@ describe('.upsertRecord', () => {
         "{title} = 'test-create'",
       ),
     ).rejects.toThrowError('Argument "record" is required');
-    done();
   });
 
-  test('should throw an error if you pass a field that does not exist', async (done) => {
+  test('should throw an error if you pass a field that does not exist', async () => {
     await expect(
       asyncAirtable.upsertRecord(
         process.env.AIRTABLE_TABLE || '',
@@ -108,10 +101,9 @@ describe('.upsertRecord', () => {
         },
       ),
     ).rejects.toThrowError(/UNKNOWN_FIELD_NAME/g);
-    done();
   });
 
-  test('should throw an error if pass a field with the incorrect data type', async (done) => {
+  test('should throw an error if pass a field with the incorrect data type', async () => {
     await expect(
       asyncAirtable.upsertRecord(
         process.env.AIRTABLE_TABLE || '',
@@ -122,10 +114,9 @@ describe('.upsertRecord', () => {
         },
       ),
     ).rejects.toThrowError(/INVALID_VALUE_FOR_COLUMN/g);
-    done();
   });
 
-  test('should throw an error if pass the table argument with an incorrect data type', async (done) => {
+  test('should throw an error if pass the table argument with an incorrect data type', async () => {
     await expect(
       asyncAirtable.upsertRecord(
         // @ts-ignore
@@ -134,10 +125,9 @@ describe('.upsertRecord', () => {
         JSON.parse(process.env.UPDATE_RECORD || ''),
       ),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
   });
 
-  test('should throw an error if pass the filterString argument with an incorrect data type', async (done) => {
+  test('should throw an error if pass the filterString argument with an incorrect data type', async () => {
     await expect(
       asyncAirtable.upsertRecord(
         process.env.AIRTABLE_TABLE || '',
@@ -146,10 +136,9 @@ describe('.upsertRecord', () => {
         JSON.parse(process.env.UPDATE_RECORD || ''),
       ),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
   });
 
-  test('should throw an error if pass the record argument with an incorrect data type', async (done) => {
+  test('should throw an error if pass the record argument with an incorrect data type', async () => {
     await expect(
       asyncAirtable.upsertRecord(
         process.env.AIRTABLE_TABLE || '',
@@ -158,31 +147,5 @@ describe('.upsertRecord', () => {
         10,
       ),
     ).rejects.toThrowError(/Incorrect data type/g);
-    done();
-  });
-
-  test('should if rate limited', async (done) => {
-    let results = [];
-    for (let i = 0; i < parseInt(process.env.REQ_COUNT || ''); i++) {
-      results.push(
-        asyncAirtable.upsertRecord(
-          process.env.AIRTABLE_TABLE || '',
-          "{title} = 'test-upsert-2'",
-          JSON.parse(process.env.UPDATE_RECORD || ''),
-        ),
-      );
-    }
-    const data: AirtableRecord[] = await Promise.all(results);
-    data.forEach((result) => {
-      expect(result).toBeDefined();
-      expect(typeof result).toBe('object');
-      expect(Object.keys(result).length).toBeGreaterThan(0);
-      expect(result.id).toBeDefined();
-      expect(result.fields).toBeDefined();
-      expect(result.createdTime).toBeDefined();
-      expect(Object.keys(result.fields).length).toBeGreaterThan(0);
-      expect(JSON.stringify(result)).not.toEqual(JSON.stringify(initResult[0]));
-    });
-    done();
   });
 });
