@@ -1,8 +1,8 @@
-// For a detailed explanation regarding each configuration property, visit:
-// https://jestjs.io/docs/en/configuration.html
 require('dotenv').config();
 
 module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
   globalSetup: './src/tests/globalSetup.ts',
   globalTeardown: './src/tests/globalTeardown.ts',
   verbose: true,
@@ -11,5 +11,9 @@ module.exports = {
   testEnvironment: 'node',
   collectCoverage: true,
   coverageDirectory: 'coverage',
-  collectCoverageFrom: ['src/**/*.{js,ts}', '!<rootDir>/node_modules/'],
+  collectCoverageFrom: [
+    'src/**/*.{js,ts}',
+    '!<rootDir>/node_modules/',
+    '!src/types/**/*',
+  ],
 };
