@@ -1,9 +1,5 @@
-import fetch from './fetch';
 import buildOpts from './buildOpts';
-import checkError from './checkError';
 import checkArg from './checkArg';
-import rateLimitHandler from './rateLimitHandler';
-const baseURL = 'https://api.airtable.com/v0';
 import {
   SelectOptions,
   AirtableDeletedResponse,
@@ -17,7 +13,8 @@ import {
   Typecast,
   updateOpts,
   bulkQueryBody,
-} from './@types';
+} from './types';
+import { request } from './http';
 
 /** @ignore */
 declare global {
@@ -39,7 +36,7 @@ const validOptions: string[] = [
 /**
  * The main AsyncAirtable class.
  */
-class AsyncAirtable {
+export class AsyncAirtable {
   /**
    * @default=true
    * This decides whether or not the library will
@@ -62,6 +59,8 @@ class AsyncAirtable {
   apiKey: string;
   /** The base id from AirTable */
   base: string;
+  /** the base URL to use when making API requests */
+  baseURL: string;
 
   /**
    * Creates a new instance of the AsyncAirtable library.
@@ -77,6 +76,7 @@ class AsyncAirtable {
     this.retryOnRateLimit = config?.retryOnRateLimit || true;
     this.retryTimeout = config?.retryTimeout || 5000;
     this.maxRetry = config?.maxRetry || 60000;
+    this.baseURL = config?.baseURL || 'https://api.airtable.com/v0';
   }
 
   /**
@@ -96,7 +96,7 @@ class AsyncAirtable {
       checkArg(table, 'table', 'string');
       checkArg(options, 'options', 'object', false);
       checkArg(page, 'page', 'number', false);
-      const url = `${baseURL}/${this.base}/${table}`;
+      const url = `${this.baseURL}/${this.base}/${table}`;
       const opts: SelectOptions = options ? { ...options } : {};
       Object.keys(opts).forEach((option) => {
         if (!validOptions.includes(option)) {
@@ -111,29 +111,15 @@ class AsyncAirtable {
             opts.offset = offset;
           }
           try {
-            const res: Response = await fetch(`${url}?${buildOpts(opts)}`, {
-              headers: { Authorization: `Bearer ${this.apiKey}` },
+            const body = await request<AirtableRecordResponse>({
+              endpoint: `${url}?${buildOpts(opts)}`,
+              options: { headers: { Authorization: `Bearer ${this.apiKey}` } },
+              instance: this,
+              pageHandler: {
+                index: i,
+                page: page,
+              },
             });
-            const body: AirtableRecordResponse = await res.json();
-            if (checkError(res.status)) {
-              if (res.status !== 429) {
-                throw new Error(JSON.stringify(body));
-              }
-
-              if (this.retryOnRateLimit) {
-                if (i + 1 === page) {
-                  return await rateLimitHandler(
-                    `${url}?${buildOpts(opts)}`,
-                    {
-                      headers: { Authorization: `Bearer ${this.apiKey}` },
-                    },
-                    this.retryTimeout,
-                    this.maxRetry,
-                    'records',
-                  );
-                }
-              }
-            }
             if (i + 1 === page) {
               return body.records;
             }
@@ -149,27 +135,13 @@ class AsyncAirtable {
             opts.offset = offset;
           }
           try {
-            const res: Response = await fetch(`${url}?${buildOpts(opts)}`, {
-              headers: { Authorization: `Bearer ${this.apiKey}` },
+            const body: AirtableRecordResponse = await request({
+              endpoint: `${url}?${buildOpts(opts)}`,
+              options: {
+                headers: { Authorization: `Bearer ${this.apiKey}` },
+              },
+              instance: this,
             });
-            const body: AirtableRecordResponse = await res.json();
-            if (checkError(res.status)) {
-              if (res.status !== 429) {
-                throw new Error(JSON.stringify(body));
-              }
-
-              if (this.retryOnRateLimit) {
-                return await rateLimitHandler(
-                  `${url}?${buildOpts(opts)}`,
-                  {
-                    headers: { Authorization: `Bearer ${this.apiKey}` },
-                  },
-                  this.retryTimeout,
-                  this.maxRetry,
-                  'records',
-                );
-              }
-            }
             data = data.concat(body.records);
             offset = body.offset;
             if (!body.offset) {
@@ -197,27 +169,14 @@ class AsyncAirtable {
     try {
       checkArg(table, 'table', 'string');
       checkArg(id, 'id', 'string');
-      const url = `${baseURL}/${this.base}/${table}/${id}`;
-      const res: Response = await fetch(url, {
-        headers: { Authorization: `Bearer ${this.apiKey}` },
+      const url = `${this.baseURL}/${this.base}/${table}/${id}`;
+      const data: AirtableRecord = await request({
+        endpoint: url,
+        options: {
+          headers: { Authorization: `Bearer ${this.apiKey}` },
+        },
+        instance: this,
       });
-      const data: AirtableRecord = await res.json();
-      if (checkError(res.status)) {
-        if (res.status !== 429) {
-          throw new Error(JSON.stringify(data));
-        }
-
-        if (this.retryOnRateLimit) {
-          return await rateLimitHandler(
-            url,
-            {
-              headers: { Authorization: `Bearer ${this.apiKey}` },
-            },
-            this.retryTimeout,
-            this.maxRetry,
-          );
-        }
-      }
       return data;
     } catch (err) {
       throw new Error(err);
@@ -241,41 +200,23 @@ class AsyncAirtable {
       checkArg(table, 'table', 'string');
       checkArg(record, 'record', 'object');
       checkArg(typecast, 'typecast', 'boolean', false);
-      const url = `${baseURL}/${this.base}/${table}`;
+      const url = `${this.baseURL}/${this.base}/${table}`;
       const body: queryBody = { fields: record };
       if (typecast !== undefined) {
         body.typecast = typecast;
       }
-      const res: Response = await fetch(url, {
-        method: 'post',
-        body: JSON.stringify(body),
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
-          'Content-Type': 'application/json',
+      const data: AirtableRecord = await request({
+        endpoint: url,
+        instance: this,
+        options: {
+          method: 'post',
+          body: JSON.stringify(body),
+          headers: {
+            Authorization: `Bearer ${this.apiKey}`,
+            'Content-Type': 'application/json',
+          },
         },
       });
-      const data: AirtableRecord = await res.json();
-      if (checkError(res.status)) {
-        if (res.status !== 429) {
-          throw new Error(JSON.stringify(data));
-        }
-
-        if (this.retryOnRateLimit) {
-          return await rateLimitHandler(
-            url,
-            {
-              method: 'post',
-              body: JSON.stringify(body),
-              headers: {
-                Authorization: `Bearer ${this.apiKey}`,
-                'Content-Type': 'application/json',
-              },
-            },
-            this.retryTimeout,
-            this.maxRetry,
-          );
-        }
-      }
       return data;
     } catch (err) {
       throw new Error(err);
@@ -302,41 +243,23 @@ class AsyncAirtable {
         checkArg(opts.destructive, 'opts.desctructive', 'boolean');
         checkArg(opts.typecast, 'opts.typecast', 'boolean', false);
       }
-      const url = `${baseURL}/${this.base}/${table}/${record.id}`;
+      const url = `${this.baseURL}/${this.base}/${table}/${record.id}`;
       const body: queryBody = { fields: record.fields };
       if (opts?.typecast !== undefined) {
         body.typecast = opts?.typecast;
       }
-      const res: Response = await fetch(url, {
-        method: opts?.destructive ? 'put' : 'patch',
-        body: JSON.stringify(body),
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
-          'Content-Type': 'application/json',
+      const data: AirtableRecord = await request({
+        endpoint: url,
+        instance: this,
+        options: {
+          method: opts?.destructive ? 'put' : 'patch',
+          body: JSON.stringify(body),
+          headers: {
+            Authorization: `Bearer ${this.apiKey}`,
+            'Content-Type': 'application/json',
+          },
         },
       });
-      const data: AirtableRecord = await res.json();
-      if (checkError(res.status)) {
-        if (res.status !== 429) {
-          throw new Error(JSON.stringify(data));
-        }
-
-        if (this.retryOnRateLimit) {
-          return await rateLimitHandler(
-            url,
-            {
-              method: opts?.destructive ? 'put' : 'patch',
-              body: JSON.stringify(body),
-              headers: {
-                Authorization: `Bearer ${this.apiKey}`,
-                'Content-Type': 'application/json',
-              },
-            },
-            this.retryTimeout,
-            this.maxRetry,
-          );
-        }
-      }
       return data;
     } catch (err) {
       throw new Error(err);
@@ -354,33 +277,17 @@ class AsyncAirtable {
     try {
       checkArg(table, 'table', 'string');
       checkArg(id, 'id', 'string');
-      const url = `${baseURL}/${this.base}/${table}/${id}`;
-      const res: Response = await fetch(url, {
-        method: 'delete',
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
+      const url = `${this.baseURL}/${this.base}/${table}/${id}`;
+      const data: DeleteResponse = await request({
+        endpoint: url,
+        instance: this,
+        options: {
+          method: 'delete',
+          headers: {
+            Authorization: `Bearer ${this.apiKey}`,
+          },
         },
       });
-      const data: DeleteResponse = await res.json();
-      if (checkError(res.status)) {
-        if (res.status !== 429) {
-          throw new Error(JSON.stringify(data));
-        }
-
-        if (this.retryOnRateLimit) {
-          return await rateLimitHandler(
-            url,
-            {
-              method: 'delete',
-              headers: {
-                Authorization: `Bearer ${this.apiKey}`,
-              },
-            },
-            this.retryTimeout,
-            this.maxRetry,
-          );
-        }
-      }
       return data;
     } catch (err) {
       throw new Error(err);
@@ -404,7 +311,7 @@ class AsyncAirtable {
       checkArg(table, 'table', 'string');
       checkArg(records, 'records', 'array');
       checkArg(typecast, 'typecast', 'boolean', false);
-      const url = `${baseURL}/${this.base}/${table}`;
+      const url = `${this.baseURL}/${this.base}/${table}`;
       const body: bulkQueryBody = {
         records: records.map((record) => ({
           fields: record,
@@ -413,37 +320,18 @@ class AsyncAirtable {
       if (typecast !== undefined) {
         body.typecast = typecast;
       }
-      const res: Response = await fetch(url, {
-        method: 'post',
-        body: JSON.stringify(body),
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
-          'Content-Type': 'application/json',
+      const data: AirtableRecordResponse = await request({
+        endpoint: url,
+        options: {
+          method: 'post',
+          body: JSON.stringify(body),
+          headers: {
+            Authorization: `Bearer ${this.apiKey}`,
+            'Content-Type': 'application/json',
+          },
         },
+        instance: this,
       });
-      const data: AirtableRecordResponse = await res.json();
-      if (checkError(res.status)) {
-        if (res.status !== 429) {
-          throw new Error(JSON.stringify(data));
-        }
-
-        if (this.retryOnRateLimit) {
-          return await rateLimitHandler(
-            url,
-            {
-              method: 'post',
-              body: JSON.stringify(body),
-              headers: {
-                Authorization: `Bearer ${this.apiKey}`,
-                'Content-Type': 'application/json',
-              },
-            },
-            this.retryTimeout,
-            this.maxRetry,
-            'records',
-          );
-        }
-      }
       return data.records;
     } catch (err) {
       throw new Error(err);
@@ -470,42 +358,23 @@ class AsyncAirtable {
         checkArg(opts.destructive, 'opts.desctructive', 'boolean', false);
         checkArg(opts.typecast, 'opts.typecast', 'boolean', false);
       }
-      const url = `${baseURL}/${this.base}/${table}`;
+      const url = `${this.baseURL}/${this.base}/${table}`;
       const body: bulkQueryBody = { records };
       if (opts?.typecast !== undefined) {
         body.typecast = opts?.typecast;
       }
-      const res: Response = await fetch(url, {
-        method: opts?.destructive ? 'put' : 'patch',
-        body: JSON.stringify(body),
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
-          'Content-Type': 'application/json',
+      const data: AirtableRecordResponse = await request({
+        endpoint: url,
+        options: {
+          method: opts?.destructive ? 'put' : 'patch',
+          body: JSON.stringify(body),
+          headers: {
+            Authorization: `Bearer ${this.apiKey}`,
+            'Content-Type': 'application/json',
+          },
         },
+        instance: this,
       });
-      const data: AirtableRecordResponse = await res.json();
-      if (checkError(res.status)) {
-        if (res.status !== 429) {
-          throw new Error(JSON.stringify(data));
-        }
-
-        if (this.retryOnRateLimit) {
-          return await rateLimitHandler(
-            url,
-            {
-              method: opts?.destructive ? 'put' : 'patch',
-              body: JSON.stringify(body),
-              headers: {
-                Authorization: `Bearer ${this.apiKey}`,
-                'Content-Type': 'application/json',
-              },
-            },
-            this.retryTimeout,
-            this.maxRetry,
-            'records',
-          );
-        }
-      }
       return data.records;
     } catch (err) {
       throw new Error(err);
@@ -534,34 +403,17 @@ class AsyncAirtable {
           query = `records[]=${id}`;
         }
       });
-      const url = `${baseURL}/${this.base}/${table}?${encodeURI(query)}`;
-      const res: Response = await fetch(url, {
-        method: 'delete',
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
+      const url = `${this.baseURL}/${this.base}/${table}?${encodeURI(query)}`;
+      const data: AirtableDeletedResponse = await request({
+        endpoint: url,
+        options: {
+          method: 'delete',
+          headers: {
+            Authorization: `Bearer ${this.apiKey}`,
+          },
         },
+        instance: this,
       });
-      const data: AirtableDeletedResponse = await res.json();
-      if (checkError(res.status)) {
-        if (res.status !== 429) {
-          throw new Error(JSON.stringify(data));
-        }
-
-        if (this.retryOnRateLimit) {
-          return await rateLimitHandler(
-            url,
-            {
-              method: 'delete',
-              headers: {
-                Authorization: `Bearer ${this.apiKey}`,
-              },
-            },
-            this.retryTimeout,
-            this.maxRetry,
-            'records',
-          );
-        }
-      }
       return data.records;
     } catch (err) {
       throw new Error(err);
@@ -606,8 +458,13 @@ class AsyncAirtable {
   };
 }
 
-export = AsyncAirtable;
-
 if (typeof window !== 'undefined') {
   window.AsyncAirtable = AsyncAirtable;
 }
+
+export const CREATED_TIME = (): string => 'CREATED_TIME()';
+export const NOW = (): string => 'NOW()';
+export const TODAY = (): string => 'TODAY()';
+export const ERROR = (): string => 'ERROR()';
+export const LAST_MODIFIED_TIME = (): string => 'LAST_MODIFIED_TIME()';
+export const RECORD_ID = (): string => 'RECORD_ID()';
