@@ -1,133 +1,128 @@
+import { QueryField, QueryObject } from './types';
+import { arrayFunctions } from './arrayFunctions';
+import { baseHandler } from './baseHandlers';
+import { handleLogicalFunc, logicalFunctions } from './logicalFunctions';
+import { logicalOperators } from './logicalOperators';
+import { handleTextFunc, textFunctions } from './textFunctions';
 import {
-  NumericalOperators,
-  QueryObject,
-  LogicalOperators,
-  QueryField,
-  BaseFieldType,
-  ComparisonObject,
-} from './@types';
+  allIndexesValid,
+  isBaseField,
+  isFunc,
+  isJoinArgs,
+  isQueryObject,
+  isQueryObjectArray,
+  isRegexArgs,
+  isRegexReplaceArgs,
+  isString,
+} from './typeCheckers';
+import { regexFunctions, regexReplaceFunction } from './regexFunctions';
+import {
+  handleNumericalFunc,
+  numericalFunctions,
+  numericOperators,
+} from './numericFunctions';
 
-const operators = ['=', '!=', '>', '>=', '<', '<='];
-
-export const isQueryObject = (item: QueryField): boolean => {
-  if (item === undefined) throw new Error('Missing Query Object');
-  const types = ['string', 'number', 'boolean'];
-  return item !== null && !types.includes(typeof item);
+export const operatorFunctions = {
+  ...logicalOperators,
+  ...numericOperators,
 };
 
-export const buildExpression = (obj: ComparisonObject, op: string): string => {
-  if (typeof obj !== 'object' || Array.isArray(obj))
-    throw new Error('Missing or Invalid Comparison Object');
-  if (typeof op !== 'string' && !operators.includes(op))
-    throw new Error('Missing or Invalid Comparison Operator');
-  const keys = Object.keys(obj);
-  const expressionMapper = (k: string, i: number) => {
-    const val = baseHandler(obj[k]);
-    return `{${k}} ${op} ${val}${i < keys.length - 1 ? ', ' : ''}`;
-  };
+export const handleError = (arg: QueryField): Error =>
+  new Error(`Invalid Query Object, ${JSON.stringify(arg)}`);
 
-  const exp = `${keys.map(expressionMapper).join('')}`;
-  return keys.length > 1 ? `AND(${exp})` : exp;
-};
+export const queryBuilder = (arg: QueryField): string => {
+  if (arg !== undefined) {
+    if (isFunc(arg) && !isBaseField(arg) && !isQueryObject(arg)) {
+      return arg();
+    }
+    if (isBaseField(arg)) {
+      return baseHandler(arg);
+    }
 
-export const numericalOperators: NumericalOperators = {
-  $gt: (val) => {
-    return buildExpression(val, '>');
-  },
-  $lt: (val) => {
-    return buildExpression(val, '<');
-  },
-  $gte: (val) => {
-    return buildExpression(val, '>=');
-  },
-  $lte: (val) => {
-    return buildExpression(val, '<=');
-  },
-  $eq: (val) => {
-    return buildExpression(val, '=');
-  },
-  $neq: (val) => {
-    return buildExpression(val, '!=');
-  },
-};
+    if (arg instanceof Array) {
+      const str = arg.map((a: QueryField) => queryBuilder(a)).join(', ');
+      return str.trim();
+    }
 
-export const logicalOperators: LogicalOperators = {
-  $not: (expression: QueryObject) => `NOT(${queryBuilder(expression)})`,
-  $and: (args: QueryObject[]) => {
-    let str = 'AND(';
-    args.forEach((arg, i) => {
-      str += queryBuilder(arg);
-      if (i < args.length - 1) {
-        str += ', ';
+    const keys = Object.keys(arg);
+    const vals = Object.values(arg);
+
+    if (
+      keys.length > 1 &&
+      allIndexesValid(vals) &&
+      isQueryObjectArray(keys.map((k, i) => ({ [k]: vals[i] })))
+    ) {
+      return logicalFunctions.$and(
+        keys.map((k, i) => ({ [k]: vals[i] })) as QueryObject & QueryObject[],
+      );
+    }
+
+    const key = keys[0];
+    if (arg[key] === undefined) {
+      throw new Error('Invalid query');
+    }
+
+    const val = arg[key];
+    if (val !== undefined) {
+      if (key === '$fieldName' && isString(val)) {
+        return `{${val}}`;
       }
-    });
-    return str + ')';
-  },
-  $or: (args: QueryObject[]) => {
-    let str = 'OR(';
-    args.forEach((arg, i) => {
-      str += queryBuilder(arg);
-      if (i < args.length - 1) {
-        str += ', ';
+
+      if (key === '$insert' && isString(val)) {
+        return val;
       }
-    });
-    return str + ')';
-  },
-};
+      if (key in logicalFunctions) {
+        return handleLogicalFunc(key, val);
+      } else if (key in textFunctions) {
+        return handleTextFunc(key, val);
+      } else if (key in numericalFunctions) {
+        return handleNumericalFunc(key, val);
+      } else if (key in regexFunctions && isRegexArgs(val)) {
+        return regexFunctions[key](val);
+      } else if (key in regexReplaceFunction && isRegexReplaceArgs(val)) {
+        return regexFunctions[key](val);
+      } else if (key in arrayFunctions) {
+        if (isJoinArgs(val)) {
+          return arrayFunctions.$arrayJoin(val.val, val.separator);
+        } else if (typeof val === 'string') {
+          return arrayFunctions[key](val);
+        }
+      } else if (isQueryObject(val)) {
+        const valKeys = Object.keys(val);
+        const subVals = Object.values(val);
 
-export const booleanHandler = (bool: boolean): string => {
-  if (typeof bool !== 'boolean') {
-    throw new Error('Invalid type');
-  }
-  return bool ? 'TRUE()' : 'FALSE()';
-};
+        if (
+          valKeys.length > 1 &&
+          allIndexesValid(subVals) &&
+          isQueryObjectArray(valKeys.map((k, i) => ({ [k]: subVals[i] }))) &&
+          subVals.every((v) => isQueryObject(v) || isBaseField(v))
+        ) {
+          if (valKeys.every((k) => k in logicalOperators)) {
+            return logicalFunctions.$and(
+              valKeys.map((k, i) => ({
+                [key]: { [k]: subVals[i] },
+              })),
+            );
+          }
+          if (valKeys.every((k) => k in numericOperators)) {
+            return '';
+          }
+        }
 
-export const baseHandler = (val: string | number | boolean | null): string => {
-  if (val === null) {
-    return 'BLANK()';
-  }
-  switch (typeof val) {
-    case 'number':
-      return `${val}`;
-    case 'string':
-      return `'${val}'`;
-    case 'boolean':
-      return booleanHandler(val);
-    default:
-      throw new Error('Wrong Type');
-  }
-};
-
-const queryBuilder = (query: QueryObject): string => {
-  let formulaString = '';
-  if (Object.keys(query as Record<string, QueryField>).length > 1) {
-    formulaString += logicalOperators.$and(
-      Object.keys(query).map((k) => ({ [k]: query[k] })),
-    );
-  } else {
-    for (const key in query) {
-      const current = query[key];
-
-      if (key in numericalOperators && isQueryObject(current as QueryObject)) {
-        formulaString += numericalOperators[key](
-          current as Record<string, number>,
-        );
-      } else if (
-        key in logicalOperators &&
-        (isQueryObject(current as QueryObject) || Array.isArray(current))
-      ) {
-        formulaString += logicalOperators[key](
-          current as QueryObject & QueryObject[],
-        );
-      } else {
-        formulaString += buildExpression(
-          query as Record<string, BaseFieldType>,
-          '=',
-        );
+        const valKey = valKeys[0];
+        const subVal = subVals[0];
+        if (
+          valKey in operatorFunctions &&
+          (isQueryObject(subVal) || isBaseField(subVal))
+        ) {
+          return operatorFunctions[valKey]({
+            [key]: subVal,
+          });
+        }
+      } else if (isQueryObject(val) || isBaseField(val)) {
+        return operatorFunctions.$eq({ [key]: val });
       }
     }
   }
-  return formulaString;
+  throw handleError(arg);
 };
-
-export default queryBuilder;
